Use `script.common.plugin.cache` correctly
It turns out I misunderstood the documentation for the addon, causing me to use
it wrongly, which meant responses where cached forever instead of only for an
hour which was the intention. The broken cache usage has been included in the
addon since 2017-01-08, meaning it's been broken for over 8 months!

This is sad :(